split take_random module and allow inlining of TakeRandom impls
