+++ polars/polars-core/src/chunked_array/ops/take.rs
@@ -13,16 +13,11 @@ use crate::chunked_array::kernels::take::{
     take_primitive_opt_iter_unchecked, take_utf8, take_utf8_iter, take_utf8_iter_unchecked,
     take_utf8_opt_iter_unchecked,
 };
-use crate::chunked_array::ops::downcast::Chunks;
 use crate::prelude::*;
 use crate::utils::NoNull;
-use arrow::array::{
-    Array, ArrayRef, BooleanArray, LargeListArray, LargeStringArray, PrimitiveArray,
-};
+use arrow::array::{Array, ArrayRef};
 use arrow::compute::kernels::take::take;
-use std::convert::TryFrom;
 use std::ops::Deref;
-use unsafe_unwrap::UnsafeUnwrap;
 
 macro_rules! take_iter_n_chunks {
     ($ca:expr, $indices:expr) => {{
@@ -547,331 +542,6 @@ pub trait AsTakeIndex {
     }
 
     fn take_index_len(&self) -> usize;
-}
-
-/// Create a type that implements a faster `TakeRandom`.
-pub trait IntoTakeRandom<'a> {
-    type Item;
-    type TakeRandom;
-    /// Create a type that implements `TakeRandom`.
-    fn take_rand(&self) -> Self::TakeRandom;
-}
-
-/// Choose the Struct for multiple chunks or the struct for a single chunk.
-macro_rules! many_or_single {
-    ($self:ident, $StructSingle:ident, $StructMany:ident) => {{
-        if $self.chunks.len() == 1 {
-            Box::new($StructSingle {
-                arr: $self.downcast_iter().next().unwrap(),
-            })
-        } else {
-            Box::new($StructMany {
-                ca: $self,
-                chunks: $self.downcast_chunks(),
-            })
-        }
-    }};
-}
-
-impl<'a, T> IntoTakeRandom<'a> for &'a ChunkedArray<T>
-where
-    T: PolarsNumericType,
-{
-    type Item = T::Native;
-    type TakeRandom = Box<dyn TakeRandom<Item = Self::Item> + 'a>;
-
-    fn take_rand(&self) -> Self::TakeRandom {
-        match self.cont_slice() {
-            Ok(slice) => Box::new(NumTakeRandomCont { slice }),
-            _ => {
-                let mut chunks = self.downcast_iter();
-                if self.chunks.len() == 1 {
-                    Box::new(NumTakeRandomSingleChunk {
-                        arr: chunks.next().unwrap(),
-                    })
-                } else {
-                    Box::new(NumTakeRandomChunked {
-                        ca: self,
-                        chunks: chunks.collect(),
-                    })
-                }
-            }
-        }
-    }
-}
-
-impl<'a> IntoTakeRandom<'a> for &'a Utf8Chunked {
-    type Item = &'a str;
-    type TakeRandom = Box<dyn TakeRandom<Item = Self::Item> + 'a>;
-
-    fn take_rand(&self) -> Self::TakeRandom {
-        many_or_single!(self, Utf8TakeRandomSingleChunk, Utf8TakeRandom)
-    }
-}
-
-impl<'a> IntoTakeRandom<'a> for &'a BooleanChunked {
-    type Item = bool;
-    type TakeRandom = Box<dyn TakeRandom<Item = Self::Item> + 'a>;
-
-    fn take_rand(&self) -> Self::TakeRandom {
-        many_or_single!(self, BoolTakeRandomSingleChunk, BoolTakeRandom)
-    }
-}
-
-impl<'a> IntoTakeRandom<'a> for &'a ListChunked {
-    type Item = Series;
-    type TakeRandom = Box<dyn TakeRandom<Item = Self::Item> + 'a>;
-
-    fn take_rand(&self) -> Self::TakeRandom {
-        let mut chunks = self.downcast_iter();
-        if self.chunks.len() == 1 {
-            Box::new(ListTakeRandomSingleChunk {
-                arr: chunks.next().unwrap(),
-                name: self.name(),
-            })
-        } else {
-            Box::new(ListTakeRandom {
-                ca: self,
-                chunks: chunks.collect(),
-            })
-        }
-    }
-}
-
-pub struct NumTakeRandomChunked<'a, T>
-where
-    T: PolarsNumericType,
-{
-    ca: &'a ChunkedArray<T>,
-    chunks: Vec<&'a PrimitiveArray<T>>,
-}
-
-macro_rules! take_random_get {
-    ($self:ident, $index:ident) => {{
-        let (chunk_idx, arr_idx) = $self.ca.index_to_chunked_index($index);
-        let arr = $self.chunks.get(chunk_idx);
-        match arr {
-            Some(arr) => {
-                if arr.is_null(arr_idx) {
-                    None
-                } else {
-                    // SAFETY:
-                    // bounds checked above
-                    unsafe { Some(arr.value_unchecked(arr_idx)) }
-                }
-            }
-            None => None,
-        }
-    }};
-}
-
-macro_rules! take_random_get_unchecked {
-    ($self:ident, $index:ident) => {{
-        let (chunk_idx, arr_idx) = $self.ca.index_to_chunked_index($index);
-        $self
-            .chunks
-            .get_unchecked(chunk_idx)
-            .value_unchecked(arr_idx)
-    }};
-}
-
-macro_rules! take_random_get_single {
-    ($self:ident, $index:ident) => {{
-        if $self.arr.is_null($index) {
-            None
-        } else {
-            // Safety:
-            // bound checked above
-            unsafe { Some($self.arr.value_unchecked($index)) }
-        }
-    }};
-}
-
-impl<'a, T> TakeRandom for NumTakeRandomChunked<'a, T>
-where
-    T: PolarsNumericType,
-{
-    type Item = T::Native;
-
-    #[inline]
-    fn get(&self, index: usize) -> Option<Self::Item> {
-        take_random_get!(self, index)
-    }
-
-    #[inline]
-    unsafe fn get_unchecked(&self, index: usize) -> Self::Item {
-        take_random_get_unchecked!(self, index)
-    }
-}
-
-pub struct NumTakeRandomCont<'a, T> {
-    slice: &'a [T],
-}
-
-impl<'a, T> TakeRandom for NumTakeRandomCont<'a, T>
-where
-    T: Copy,
-{
-    type Item = T;
-
-    #[inline]
-    fn get(&self, index: usize) -> Option<Self::Item> {
-        self.slice.get(index).copied()
-    }
-
-    #[inline]
-    unsafe fn get_unchecked(&self, index: usize) -> Self::Item {
-        *self.slice.get_unchecked(index)
-    }
-}
-
-pub struct NumTakeRandomSingleChunk<'a, T>
-where
-    T: PolarsNumericType,
-{
-    arr: &'a PrimitiveArray<T>,
-}
-
-impl<'a, T> TakeRandom for NumTakeRandomSingleChunk<'a, T>
-where
-    T: PolarsNumericType,
-{
-    type Item = T::Native;
-
-    #[inline]
-    fn get(&self, index: usize) -> Option<Self::Item> {
-        take_random_get_single!(self, index)
-    }
-
-    #[inline]
-    unsafe fn get_unchecked(&self, index: usize) -> Self::Item {
-        self.arr.value_unchecked(index)
-    }
-}
-
-pub struct Utf8TakeRandom<'a> {
-    ca: &'a Utf8Chunked,
-    chunks: Chunks<'a, LargeStringArray>,
-}
-
-impl<'a> TakeRandom for Utf8TakeRandom<'a> {
-    type Item = &'a str;
-
-    #[inline]
-    fn get(&self, index: usize) -> Option<Self::Item> {
-        take_random_get!(self, index)
-    }
-
-    #[inline]
-    unsafe fn get_unchecked(&self, index: usize) -> Self::Item {
-        let (chunk_idx, arr_idx) = self.ca.index_to_chunked_index(index);
-        self.chunks
-            .get_unchecked(chunk_idx)
-            .value_unchecked(arr_idx)
-    }
-}
-
-pub struct Utf8TakeRandomSingleChunk<'a> {
-    arr: &'a LargeStringArray,
-}
-
-impl<'a> TakeRandom for Utf8TakeRandomSingleChunk<'a> {
-    type Item = &'a str;
-
-    #[inline]
-    fn get(&self, index: usize) -> Option<Self::Item> {
-        take_random_get_single!(self, index)
-    }
-
-    #[inline]
-    unsafe fn get_unchecked(&self, index: usize) -> Self::Item {
-        self.arr.value_unchecked(index)
-    }
-}
-
-pub struct BoolTakeRandom<'a> {
-    ca: &'a BooleanChunked,
-    chunks: Chunks<'a, BooleanArray>,
-}
-
-impl<'a> TakeRandom for BoolTakeRandom<'a> {
-    type Item = bool;
-
-    #[inline]
-    fn get(&self, index: usize) -> Option<Self::Item> {
-        take_random_get!(self, index)
-    }
-
-    #[inline]
-    unsafe fn get_unchecked(&self, index: usize) -> Self::Item {
-        take_random_get_unchecked!(self, index)
-    }
-}
-
-pub struct BoolTakeRandomSingleChunk<'a> {
-    arr: &'a BooleanArray,
-}
-
-impl<'a> TakeRandom for BoolTakeRandomSingleChunk<'a> {
-    type Item = bool;
-
-    #[inline]
-    fn get(&self, index: usize) -> Option<Self::Item> {
-        take_random_get_single!(self, index)
-    }
-
-    #[inline]
-    unsafe fn get_unchecked(&self, index: usize) -> Self::Item {
-        self.arr.value(index)
-    }
-}
-pub struct ListTakeRandom<'a> {
-    ca: &'a ListChunked,
-    chunks: Vec<&'a LargeListArray>,
-}
-
-impl<'a> TakeRandom for ListTakeRandom<'a> {
-    type Item = Series;
-
-    #[inline]
-    fn get(&self, index: usize) -> Option<Self::Item> {
-        let v = take_random_get!(self, index);
-        v.map(|v| {
-            let s = Series::try_from((self.ca.name(), v));
-            s.unwrap()
-        })
-    }
-
-    #[inline]
-    unsafe fn get_unchecked(&self, index: usize) -> Self::Item {
-        let v = take_random_get_unchecked!(self, index);
-        let s = Series::try_from((self.ca.name(), v));
-        s.unwrap()
-    }
-}
-
-pub struct ListTakeRandomSingleChunk<'a> {
-    arr: &'a LargeListArray,
-    name: &'a str,
-}
-
-impl<'a> TakeRandom for ListTakeRandomSingleChunk<'a> {
-    type Item = Series;
-
-    #[inline]
-    fn get(&self, index: usize) -> Option<Self::Item> {
-        let v = take_random_get_single!(self, index);
-        v.map(|v| {
-            let s = Series::try_from((self.name, v));
-            s.unwrap()
-        })
-    }
-
-    #[inline]
-    unsafe fn get_unchecked(&self, index: usize) -> Self::Item {
-        let s = Series::try_from((self.name, self.arr.value_unchecked(index)));
-        s.unsafe_unwrap()
-    }
 }
 
 impl<T> ChunkTakeEvery<T> for ChunkedArray<T>
